test: rename some fixture variables
